+++ app/src/main/java/com/spoony/spoony/domain/repository/RegisterRepository.kt
@@ -8,12 +8,12 @@ interface RegisterRepository {
     suspend fun getFoodCategories(): Result<List<CategoryEntity>>
     suspend fun searchPlace(query: String, display: Int = 5): Result<List<PlaceEntity>>
     suspend fun checkDuplicatePlace(
-        userId: Long,
+        userId: Int,
         latitude: Double,
         longitude: Double
     ): Result<Boolean>
     suspend fun registerPost(
-        userId: Long,
+        userId: Int,
         title: String,
         description: String,
         placeName: String,
